# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Notifier API",
        "title": "Notifier API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/v1/admin/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all payments. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get all payments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PaymentResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/admin/payments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single payment by ID. Users can only access their own payments.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get payment by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payment ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/admin/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single user by their ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserPayload"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a user account by ID.\nPossible 400 errors: invalid id, invalid user data.\nPossible 401 errors: userID not in context, role not in context.\nPossible 403 errors: cannot delete another user's account.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid id or user data",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/payments/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all payments for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get my payments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PaymentResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/payments/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a payment by ID. Users can only update their own payments.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Update a payment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Validation error or invalid id",
                        "schema": {
                            "$ref": "#/definitions/dto.PaymentValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a payment by ID. Admins can delete any payment, regular users can only delete their own.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Delete a payment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid payment ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "dto.PaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 14.99
                },
                "category": {
                    "enum": [
                        "electronics",
                        "entertainment",
                        "education",
                        "clothing",
                        "work",
                        "sports"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.CategoryType"
                        }
                    ],
                    "example": "work"
                },
                "date": {
                    "type": "string",
                    "example": "2026-04-19"
                },
                "due_date": {
                    "type": "string",
                    "example": "2026-04-25"
                },
                "frequency": {
                    "enum": [
                        "daily",
                        "weekly",
                        "montly",
                        "yearly"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.FrequencyType"
                        }
                    ],
                    "example": "montly"
                },
                "id": {
                    "type": "integer",
                    "example": 7
                },
                "name": {
                    "type": "string",
                    "example": "claude"
                },
                "paid": {
                    "type": "boolean",
                    "example": false
                },
                "paid_at": {
                    "type": "string",
                    "example": "2026-04-19"
                },
                "receipt_url": {
                    "type": "string",
                    "example": "https://s3.amazonaws.com/receipts/abc123.pdf"
                },
                "recurrent": {
                    "type": "boolean",
                    "example": true
                },
                "type": {
                    "enum": [
                        "income",
                        "expense",
                        "subscription"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.TransactionType"
                        }
                    ]
                }
            }
        },
        "dto.PaymentValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid category"
                },
                "meta": {
                    "$ref": "#/definitions/dto.ValidationMeta"
                }
            }
        },
...
        "dto.UpdatePaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/entity.CategoryType"
                },
                "date": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/entity.FrequencyType"
                },
                "name": {
                    "type": "string"
                },
                "paid": {
                    "type": "boolean"
                },
                "paid_at": {
                    "type": "string"
                },
                "receipt_url": {
                    "type": "string"
                },
                "recurrent": {
                    "type": "boolean"
                },
                "type": {
                    "$ref": "#/definitions/entity.TransactionType"
                }
            }
        },
...
        "dto.UserPayload": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": false
                },
                "email": {
                    "type": "string",
                    "example": "rober@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "rober"
                }
            }
        },
...
        "dto.ValidationMeta": {
            "type": "object",
            "properties": {
                "allowed_categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CategoryType"
                    },
                    "example": [
                        "electronics",
                        "entertainment",
                        "education",
                        "clothing",
                        "work",
                        "sports"
                    ]
                },
                "allowed_frequencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FrequencyType"
                    },
                    "example": [
                        "daily",
                        "weekly",
                        "monthly",
                        "yearly"
                    ]
                },
                "allowed_types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.TransactionType"
                    },
                    "example": [
                        "expense",
                        "income",
                        "subscription"
                    ]
                }
            }
        },
...
        "entity.CategoryType": {
            "type": "string",
            "enum": [
                "electronics",
                "entertainment",
                "education",
                "clothing",
                "work",
                "sports"
            ],
            "x-enum-varnames": [
                "CategoryTypeElectronics",
                "CategoryTypeEntertainment",
                "CategoryTypeEducation",
                "CategoryTypeClothing",
                "CategoryTypeWork",
                "CategoryTypeSports"
            ]
        },
        "entity.FrequencyType": {
            "type": "string",
            "enum": [
                "daily",
                "weekly",
                "monthly",
                "yearly"
            ],
            "x-enum-varnames": [
                "FrequencyTypeDaily",
                "FrequencyTypeWeekly",
                "FrequencyTypeMonthly",
                "FrequencyTypeYearly"
            ]
        },
        "entity.TransactionType": {
            "type": "string",
            "enum": [
                "expense",
                "income",
                "subscription"
            ],
            "x-enum-varnames": [
                "TransactionTypeExpense",
                "TransactionTypeIncome",
                "TransactionTypeSubscription"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}